better folder organize and set dev prod mark for meddleware
